check-point-1

+++ repository/contacts.js
@@ -1,29 +1,38 @@
 const Contact = require('../models/contact')
 
-const listContacts = async () => {
-	const result = await Contact.find()
-	return result
+const listContacts = async ({ limit, skip, sortCriteria, select }, user) => {
+	const { docs: contacts, ...rest } = await Contact.paginate(
+		{ owner: user.id },
+		{ limit, offset: skip, sort: sortCriteria, select },
+	)
+	return { ...contacts, ...rest }
 }
-const getContactById = async (contactId) => {
-	const result = await Contact.findOne({ _id: contactId })
+
+const getContactById = async (contactId, user) => {
+	const result = await Contact.findOne({ _id: contactId, owner: user._id })
+		.populate({
+			path: 'owner',
+			select: 'name email subscription',
+		})
 	return result
 }
 
-const removeContact = async (contactId) => {
-	const result = await Contact.findOneAndRemove({ _id: contactId })
+const removeContact = async (contactId, user) => {
+	const result = await Contact.findOneAndRemove({ _id: contactId, owner: user._id })
 	return result
 }
 
-const addContact = async (body) => {
-	const result = await Contact.create(body)
+const addContact = async (body, user) => {
+	const result = await Contact.create({ ...body, owner: user._id })
 	return result
 }
 
-const updateContact = async (contactId, body) => {
+const updateContact = async (contactId, body, user) => {
 	const result = await Contact.findOneAndUpdate(
-		{ _id: contactId },
+		{ _id: contactId, owner: user._id },
 		{ ...body },
-		{ new: true })
+		{ new: true },
+	)
 	return result
 }
